+++ annotations/src/main/java/arez/annotations/Observable.java
@@ -28,7 +28,7 @@ import javax.annotation.Nonnull;
  * <li>Must not be private</li>
  * <li>Must not be static</li>
  * <li>Must not be final</li>
- * <li>Must not be abstract</li>
+ * <li>May be abstract but if abstract then the paired setter or getter must also be abstract</li>
  * </ul>
  */
 @Documented