+++ processor/src/test/resources/bad_input/com/example/observable/NonAbstractSetterModel.java
@@ -1,0 +1,19 @@
+package com.example.observable;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+
+@ArezComponent
+public abstract class NonAbstractSetterModel
+{
+  private long _field;
+
+  @Observable
+  public abstract long getField();
+
+  @Observable
+  public void setField( final long field )
+  {
+    _field = field;
+  }
+}