+++ processor/src/test/resources/expected/Arez_BasicModelWithDifferentAccessLevels.java
@@ -134,7 +134,7 @@ public final class Arez_BasicModelWithDifferentAccessLevels extends BasicModelWi
   @Override
   public void setValue(final String value) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
-    if ( !Objects.equals(value, super.getValue()) ) {
+    if ( !Objects.equals( value, super.getValue() ) ) {
       this.$$arez$$_value.preReportChanged();
       super.setValue(value);
       this.$$arez$$_value.reportChanged();