+++ processor/src/test/resources/expected/com/example/repository/Arez_CompleteRepositoryExample.java
@@ -111,7 +111,7 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
   @Override
   public void setName(@Nonnull final String name) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
-    if ( !Objects.equals(name, super.getName()) ) {
+    if ( !Objects.equals( name, super.getName() ) ) {
       this.$$arez$$_name.preReportChanged();
       super.setName(name);
       this.$$arez$$_name.reportChanged();
@@ -129,7 +129,7 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
   @Override
   public void setPackageName(@Nonnull final String packageName) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
-    if ( !Objects.equals(packageName, super.getPackageName()) ) {
+    if ( !Objects.equals( packageName, super.getPackageName() ) ) {
       this.$$arez$$_packageName.preReportChanged();
       super.setPackageName(packageName);
       this.$$arez$$_packageName.reportChanged();
@@ -147,7 +147,7 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
   @Override
   public void setQualifiedName(@Nullable final String qualifiedName) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
-    if ( !Objects.equals(qualifiedName, super.getRawQualifiedName()) ) {
+    if ( !Objects.equals( qualifiedName, super.getRawQualifiedName() ) ) {
       this.$$arez$$_rawQualifiedName.preReportChanged();
       super.setQualifiedName(qualifiedName);
       this.$$arez$$_rawQualifiedName.reportChanged();