+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleCtors.java
@@ -127,7 +127,7 @@ final class Arez_RepositoryWithMultipleCtors extends RepositoryWithMultipleCtors
   @Override
   public void setName(@Nonnull final String name) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
-    if ( !Objects.equals(name, super.getName()) ) {
+    if ( !Objects.equals( name, super.getName() ) ) {
       this.$$arez$$_name.preReportChanged();
       super.setName(name);
       this.$$arez$$_name.reportChanged();